added a preventative guard for the Completed checkbox, so if a user presses it more than once very quickly, it would not process the second request until the first request finishes. Also moved the function, which does this action, into its own file because the code was almost identical in both the Reminders and Search files.
