+++ client/src/components/dashboard/reminders/Reminders.js
@@ -6,7 +6,7 @@ import { Link } from "react-router-dom";
 // Components:
 import CreateReminder from "./CreateReminder";
 import EditReminder from "./EditReminder";
-import { PushGeneralReminderTimeAhead } from "./PushGeneralReminderTimeAhead";
+import { ReminderCompletedState } from "./ReminderCompletedState";
 
 import "./../../../App.css";
 import "react-tabs/style/react-tabs.css";
@@ -17,6 +17,8 @@ function Reminders({ isAuth }) {
 	const [allCompletedReminders, setAllCompletedReminders] = useState([]);
 	const [allOverdueReminders, setAllOverdueReminders] = useState([]);
 	const [allReminders, setAllReminders] = useState([]);
+
+	var completedComputationInProgress = false;
 
 	const textValues = ["Delete", "Are you sure?", "Deleting..."];
 	var isExecuting = false;
@@ -128,115 +130,21 @@ function Reminders({ isAuth }) {
 		}
 	}
 
-	async function reminderCompletedState(reminder_id, reminder_completed) {
-		reminder_completed = !reminder_completed;
+	async function reminderCompletedStateHelper(reminder_id, reminder_completed) {
 
-		const myHeaders = new Headers();
-		myHeaders.append("Content-type", "application/json");
-		myHeaders.append("token", localStorage.token);
+		/* Preventative flag so if a user quickly clicks on the completed box, before the 
+		 * computation finishes, the requests reach the database and the second request
+		 * fails because it doesn't have the proper data anymore.
+		 */
+		if (!completedComputationInProgress) {
+			completedComputationInProgress = true;
+			let activeRemindersEmpty = allActiveReminders.length === 0;
+		
+			await ReminderCompletedState(reminder_id, reminder_completed, activeRemindersEmpty);
 
-		try {
-			if (reminder_completed) {
-				// The Completed checkbox just got checked.
-	
-				const respActiveReminders = await fetch(
-					`/api/dashboard/reminder/active/${reminder_id}`, {
-						method: "DELETE",
-						headers: {token: localStorage.token}
-					}
-				);
-	
-				// eslint-disable-next-line
-				const respOverdueReminders = await fetch(
-					`/api/dashboard/reminder/overdue/${reminder_id}`, {
-						method: "DELETE",
-						headers: {token: localStorage.token}
-					}
-				);
-
-				const parseResp = await respActiveReminders.json();
-				
-				const id = reminder_id;
-				const completed = reminder_completed,
-					title = parseResp.reminder_title,
-					desc = parseResp.reminder_desc,
-					dueDate = parseResp.reminder_due_date,
-					reminderDate = parseResp.reminder_reminder_date,
-					reminderSent = parseResp.reminder_reminder_sent;
-	
-				const body = { completed, title, desc, dueDate, reminderDate, reminderSent };
-				const bodyPlusId = { id, completed, title, desc, dueDate, reminderDate, reminderSent };
-	
-				// eslint-disable-next-line
-				const respAllReminders = await fetch(
-					`/api/dashboard/reminder/all/${reminder_id}`,
-					{
-						method: "PUT",
-						headers: myHeaders,
-						body: JSON.stringify(body)
-					}
-				);
-	
-				// eslint-disable-next-line
-				const respCompletedReminders = await fetch(
-					"/api/dashboard/reminder/completed", {
-						method: "POST",
-						headers: myHeaders,
-						body: JSON.stringify(bodyPlusId)
-				});
-	
-			} else {
-				// The Completed checkbox has now been unchecked.
-
-				if (allActiveReminders.length === 0) {
-					await PushGeneralReminderTimeAhead(myHeaders);
-				}
-	
-				// eslint-disable-next-line
-				const respCompletedReminders = await fetch(
-					`/api/dashboard/reminder/completed/${reminder_id}`,
-					{
-						method: "DELETE",
-						headers: {token: localStorage.token}
-					}
-				);
-	
-				const parseResp = await respCompletedReminders.json();
-	
-				const id = reminder_id;
-				const completed = reminder_completed,
-					title = parseResp.reminder_title,
-					desc = parseResp.reminder_desc,
-					dueDate = parseResp.reminder_due_date,
-					reminderDate = parseResp.reminder_reminder_date,
-					reminderSent = parseResp.reminder_reminder_sent;
-	
-				const body = { completed, title, desc, dueDate, reminderDate, reminderSent };
-				const bodyPlusId = { id, completed, title, desc, dueDate, reminderDate, reminderSent };
-	
-				// eslint-disable-next-line
-				const respAllReminders = await fetch(
-					`/api/dashboard/reminder/all/${reminder_id}`,
-					{
-						method: "PUT",
-						headers: myHeaders,
-						body: JSON.stringify(body)
-					}
-				);
-	
-				// eslint-disable-next-line
-				const respActiveReminders = await fetch(
-					"/api/dashboard/reminder/active", {
-						method: "POST",
-						headers: myHeaders,
-						body: JSON.stringify(bodyPlusId)
-				});
-			}
-	
+			// Reloading is necessary because the Edit component will not have the
+			// same state without a rerender.
 			window.location = "/";
-
-		} catch (error) {
-			console.error(error.message);
 		}
 	}
 
@@ -405,7 +313,7 @@ function Reminders({ isAuth }) {
 												type="checkbox"
 												className="completed-checkboxes"
 												onChange={() =>
-													reminderCompletedState(
+													reminderCompletedStateHelper(
 														currReminder.reminder_id,
 														currReminder.reminder_completed
 													)
@@ -464,7 +372,7 @@ function Reminders({ isAuth }) {
 												type="checkbox"
 												className="completed-checkboxes"
 												onChange={() =>
-													reminderCompletedState(
+													reminderCompletedStateHelper(
 														currReminder.reminder_id,
 														currReminder.reminder_completed
 													)
@@ -523,7 +431,7 @@ function Reminders({ isAuth }) {
 												type="checkbox"
 												className="completed-checkboxes"
 												onChange={() =>
-													reminderCompletedState(
+													reminderCompletedStateHelper(
 														currReminder.reminder_id,
 														currReminder.reminder_completed
 													)
@@ -583,7 +491,7 @@ function Reminders({ isAuth }) {
 												type="checkbox"
 												className="completed-checkboxes"
 												onChange={() =>
-													reminderCompletedState(
+													reminderCompletedStateHelper(
 														currReminder.reminder_id,
 														currReminder.reminder_completed
 													)
