+++ client/src/components/dashboard/reminders/Search.js
@@ -2,7 +2,7 @@ import React, { Fragment, useState, useEffect } from 'react';
 import InlineConfirmButton from "react-inline-confirm";
 
 import EditReminder from "./EditReminder";
-import { PushGeneralReminderTimeAhead } from "./PushGeneralReminderTimeAhead";
+import { ReminderCompletedState } from "./ReminderCompletedState";
 
 import "./../../../App.css";
 
@@ -12,6 +12,8 @@ function Search({ isAuth }) {
     const [filteredReminders, setFilteredReminders] = useState([]);
     const [titleSearched, setTitleSearched] = useState("");
 	const [activeRemindersEmpty, setActiveRemindersEmpty] = useState(true);
+
+	var completedComputationInProgress = false;
 
     const textValues = ["Delete", "Are you sure?", "Deleting..."];
 	var isExecuting = false;
@@ -67,6 +69,7 @@ function Search({ isAuth }) {
         }
     }
 
+	// Action when reminder Delete button is clicked.
     async function deleteReminderTask(reminder_id) {
 		try {
 			// eslint-disable-next-line
@@ -115,115 +118,21 @@ function Search({ isAuth }) {
 		}
     }
     
-    async function reminderCompletedState(reminder_id, reminder_completed) {
-		reminder_completed = !reminder_completed;
+	// Action when Completed checkbox is clicked.
+    async function reminderCompletedStateHelper(reminder_id, reminder_completed) {
+		
+		/* Preventative flag so if a user quickly clicks on the completed box, before the 
+		 * computation finishes, the requests reach the database and the second request
+		 * fails because it doesn't have the proper data anymore.
+		 */
+		if (!completedComputationInProgress) {
+			completedComputationInProgress = true;
 
-		const myHeaders = new Headers();
-		myHeaders.append("Content-type", "application/json");
-		myHeaders.append("token", localStorage.token);
+			await ReminderCompletedState(reminder_id, reminder_completed, activeRemindersEmpty);
 
-		try {
-			if (reminder_completed) {
-				// The Completed checkbox just got checked.
-	
-				const respActiveReminders = await fetch(
-					`/api/dashboard/reminder/active/${reminder_id}`, {
-						method: "DELETE",
-						headers: {token: localStorage.token}
-					}
-				);
-	
-				// eslint-disable-next-line
-				const respOverdueReminders = await fetch(
-					`/api/dashboard/reminder/overdue/${reminder_id}`, {
-						method: "DELETE",
-						headers: {token: localStorage.token}
-					}
-				);
-
-				const parseResp = await respActiveReminders.json();
-				
-				const id = reminder_id;
-				const completed = reminder_completed,
-					title = parseResp.reminder_title,
-					desc = parseResp.reminder_desc,
-					dueDate = parseResp.reminder_due_date,
-					reminderDate = parseResp.reminder_reminder_date,
-					reminderSent = parseResp.reminder_reminder_sent;
-	
-				const body = { completed, title, desc, dueDate, reminderDate, reminderSent };
-				const bodyPlusId = { id, completed, title, desc, dueDate, reminderDate, reminderSent };
-	
-				// eslint-disable-next-line
-				const respAllReminders = await fetch(
-					`/api/dashboard/reminder/all/${reminder_id}`,
-					{
-						method: "PUT",
-						headers: myHeaders,
-						body: JSON.stringify(body)
-					}
-				);
-	
-				// eslint-disable-next-line
-				const respCompletedReminders = await fetch(
-					"/api/dashboard/reminder/completed", {
-						method: "POST",
-						headers: myHeaders,
-						body: JSON.stringify(bodyPlusId)
-				});
-	
-			} else {
-				// The Completed checkbox has now been unchecked.
-
-				if (activeRemindersEmpty) {
-					await PushGeneralReminderTimeAhead(myHeaders);
-				}
-	
-				// eslint-disable-next-line
-				const respCompletedReminders = await fetch(
-					`/api/dashboard/reminder/completed/${reminder_id}`,
-					{
-						method: "DELETE",
-						headers: {token: localStorage.token}
-					}
-				);
-	
-				const parseResp = await respCompletedReminders.json();
-	
-				const id = reminder_id;
-				const completed = reminder_completed,
-					title = parseResp.reminder_title,
-					desc = parseResp.reminder_desc,
-					dueDate = parseResp.reminder_due_date,
-					reminderDate = parseResp.reminder_reminder_date,
-					reminderSent = parseResp.reminder_reminder_sent;
-	
-				const body = { completed, title, desc, dueDate, reminderDate, reminderSent };
-				const bodyPlusId = { id, completed, title, desc, dueDate, reminderDate, reminderSent };
-	
-				// eslint-disable-next-line
-				const respAllReminders = await fetch(
-					`/api/dashboard/reminder/all/${reminder_id}`,
-					{
-						method: "PUT",
-						headers: myHeaders,
-						body: JSON.stringify(body)
-					}
-				);
-	
-				// eslint-disable-next-line
-				const respActiveReminders = await fetch(
-					"/api/dashboard/reminder/active", {
-						method: "POST",
-						headers: myHeaders,
-						body: JSON.stringify(bodyPlusId)
-				});
-			}
-	
+			// Reloading is necessary because the Edit component will not have the
+			// same state without a rerender.
 			window.location = "/Dashboard/Search";
-
-		} catch (error) {
-			console.error(error.message);
 		}
 	}
 
@@ -286,7 +195,7 @@ function Search({ isAuth }) {
                                             type="checkbox"
                                             className="completed-checkboxes"
                                             onChange={() =>
-                                                reminderCompletedState(
+                                                reminderCompletedStateHelper(
                                                     currReminder.reminder_id,
                                                     currReminder.reminder_completed
                                                 )
